Replace DOUBLE PRECISION/DOUBLE COMPLEX with KIND=dp declarations
- unifies precision declarations on REAL(KIND=dp)/COMPLEX(KIND=dp)
- folds in 4 leftover REAL*8 sites in DirectSolve.F90 and ZirkaTest.F90
- adds USE Types, ONLY : dp where a declaration lives in an INTERFACE block or standalone subroutine that doesn't get dp via host association

+++ fem/src/DirectSolve.F90
@@ -3448,7 +3448,7 @@ CONTAINS
     INTEGER, POINTER :: Iparm(:)
     INTEGER i, j, k, nz, idum(1), nzutd
     LOGICAL :: Found, matsym, matpd
-    REAL*8  :: ddum(1)
+    REAL(KIND=dp)  :: ddum(1)
 
     LOGICAL :: Factorize, FreeFactorize
     INTEGER :: tlen, allocstat
@@ -3674,7 +3674,7 @@ CONTAINS
       INTEGER, POINTER :: Iparm(:)
       INTEGER i, j, k, nz, idum(1)
       LOGICAL :: Found, Symm, Posdef
-      REAL*8  waltime1, waltime2, ddum(1), dparm(64)
+      REAL(KIND=dp)  waltime1, waltime2, ddum(1), dparm(64)
 
       LOGICAL :: Factorize, FreeFactorize
       INTEGER :: tlen
